fix bbsrf, warning free, defined HAVE_DES_CRYPT
git-svn-id: http://opensvn.csie.org/pttbbs/branches/victor.solaris@2183 63ad8ddf-47c3-0310-b6dd-a9e9d9715204

+++ util/xchatd.c
@@ -3135,7 +3135,7 @@ start_daemon()
     if (fd >= 0)
     {
 	/* sprintf(buf, "%5d\n", value); */
-	sprintf(buf, "%5d\n", getpid());
+	sprintf(buf, "%5d\n", (int)getpid());
 	write(fd, buf, 6);
 	close(fd);
     }
